data-collection/speech-first-recorder/requirements.txt
streamlit
fastapi
uvicorn[standard]
python-multipart
ffmpeg-python
PyYAML
boto3
pyAudioAnalysis
tqdm
eyed3